Proof of concept branch to investigate fixes for jsx=true warning and errors. The main issue is that in some components there is some styling declared as <style  jsx>, but when running the tests for these components we get the warning         “Warning: Received `true` for a non-boolean attribute `jsx`”.  If we change to <style  jsx = “true”>, the warning is fixed but then we get a compile error:         Type error: Type 'string' is not assignable to type 'boolean | undefined'.
I do not understand completely why not setting this variable is giving the warning, but some explanation and suggestions for fixes are provided in the following links:

https://github.com/vitejs/vite/discussions/6136
https://github.com/vercel/styled-jsx/issues/838
https://github.com/vercel/styled-jsx#rendering-in-tests

I tried all of them, but they did not fix the warning. However, in the end, by just removing the jsx tag, it seems to work ( the app compiles and not warnings).